subbrute/requirements.txt
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
configparser==3.7.3
dnspython==1.16.0
idna==2.8
lxml==4.3.1
requests==2.21.0
soupsieve==1.8
urllib3==1.24.1
